Remove customer orders list page
The page is now included in Solidus frontend.

Change users_controller to decorator file

This change allows to decorate the class `Spree::UsersController`
when already present (ie. when it's already included in the gem
`solidus_frontend`) or to define the class from scratch when it's
not (ie. when using a version of Solidus where `solidus_frontend`
does not include it).

Once all supported versions of Solidus include the account page,
we can get rid of the class definition, together with the erb
view at `lib/views/frontend/spree/users/show.html.erb`.